requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.7.0
PyQt5>=5.15.0
pyyaml>=6.0
psutil>=2.0.0
Pillow>=9.4.0
tqdm>=4.65.0
requests>=2.28.0

# Object detection
ultralytics>=8.0.0  # For YOLOv8

# Segmentation
segment-anything>=1.0  # For SAM, optional
scikit-image>=0.20.0   # For basic image processing

# LLM integration
# No specific requirements for Ollama as it's a separate application

# Optional dependencies
# tensorflow>=2.12.0  # If using TensorFlow models
# onnxruntime>=1.14.0  # If using ONNX models
# coremltools>=6.3.0   # If using CoreML on macOS
# pillow-heif>=0.5.0   # For HEIC/HEIF image format support
